Add files via upload
This is the .py file to perform an LED test along the X axis in Minecraft, The script tells the LED's to turn on after crossing a certain coordinate. There's 3 separate coordinates for 3 individual LED's. With French Commentaries. Note with this script the LED's will stay always on (until powered down).