Add files via upload
This is the .py file to perform an LED test along the X axis in Minecraft. The script tells the LED's to turn on or off after crossing a certain coordinate. There's 3 separate coordinates for 3 individual LED's. With French Commentaries. Note this adds the ability to turn off the LED's if we enter a coordinate lower then specified.