Add files via upload
This is the .py file to perform an LED test along the X and Z axes in Minecraft. The script tells the LED's to turn on or off after crossing a certain coordinate. There's 6 separate coordinates for 6 individual LED's (3 for the X axis and 3 for the Z axis). With French Commentaries, and the ability to turn off the LED's if we enter a coordinate lower then specified.